MR-62: Add Form divide and logic
